// src/app/Helper/models/Stock.ts
export class Stock {
    ProductID: number;
    ProductVariantID: number;
    Product: string;
    ProductVariant: string;
    ProductImage: string;
    RemainingStock: number;
    HoldStock?: number;
    Tax?: number;
    LuxuaryTax?: number;
    ProductVariantBarcode: string;
    IsTrackable?: boolean;
    SalePrice?: number;
    PurchasePrice?: number;
    AvgPurchasePrice?: number;
    ProductModel: string;
    ArticalNo: string;
    Barcode: string;
    Category: string;
    CategoryID?: number;
    
}       